feat : 쿠키에서 세션 ID 추출
- HttpRequest 객체에서 쿠키 정보를 파싱하고, 여기서 세션 ID (sid)를 추출
- HTTP 요청 헤더에서 Cookie 필드를 찾아 세션 ID를 찾음
