+++ src/main/java/http/request/parser/RequestCookieParser.java
@@ -1,0 +1,20 @@
+package http.request.parser;
+
+import java.util.HashMap;
+import java.util.Map;
+
+public class RequestCookieParser {
+    public static Map<String, String> parse(String cookieHeader) {
+        Map<String, String> cookies = new HashMap<>();
+        if (cookieHeader != null && !cookieHeader.isEmpty()) {
+            String[] pairs = cookieHeader.split(";");
+            for (String pair : pairs) {
+                String[] keyValue = pair.split("=", 2);
+                if (keyValue.length == 2) {
+                    cookies.put(keyValue[0].trim(), keyValue[1].trim());
+                }
+            }
+        }
+        return cookies;
+    }
+}
